requirement.txt
numpy
pandas
matplotlib.pyplot
scipy.stats
streamlit